requirements.txt
aioredis==1.3.1
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap4==3.0.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
django-phonenumber-field==5.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
install==1.3.4
Markdown==3.3.4
MarkupPy==1.14
msgpack==1.0.2
numpy==1.20.2
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.2.4
phonenumbers==8.12.21
Pillow==8.2.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
pywebio==1.2.3
PyYAML==5.4.1
reportlab==3.5.67
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.3.1
tablib==3.0.0
tornado==6.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
ua-parser==0.10.0
user-agents==2.2.0
validate-email==1.3
virtualenv==20.4.4
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
